// src/app/model/Currentuser.ts
export class Currentuser {
    userId: string;
    name: string;
    fromName: string;
    toName: string;
    online: boolean;
    image: { darkImage: string; lightImage: string; darkImageErr: string; lightImageErr: string };
    email: string;
    lastMessage: string;
    subject: string;
    role: string;
    date: Date;
    ayatollah: string;
    messageId: string;
    lastLogin: string;
    IsRead: string;
    IsReplied: string;
    showHistory:boolean;
    UserLoginStatus:string;
    userEmail:string;
    Duration:string;
}